BJS2-66437: попытался реализовать абстрактный обработчик достижений чтоб не дублировать код
