+++ src/main/java/faang/school/achievement/service/EventHandler.java
@@ -1,5 +1,5 @@
 package faang.school.achievement.service;
 
 public interface EventHandler<T> {
-    public void handle(T event);
+    void handle(T event);
 }
