+++ src/test/java/faang/school/achievement/AchievementServiceImplAppTests.java
@@ -3,7 +3,7 @@ package faang.school.achievement;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-class AchievementServiceAppTests {
+class AchievementServiceImplAppTests {
     @Test
     void contextLoads() {
         Assertions.assertThat(40 + 2).isEqualTo(42);
